+++ src/main/kotlin/pullrequestfactory/io/programs/UserProperties.kt
@@ -1,5 +1,5 @@
 package pullrequestfactory.io.programs
 
 interface UserProperties {
-    fun get_github_basic_auth_token(): String?
+    fun get_github_auth_token(): String?
 }