+++ src/main/kotlin/pullrequestfactory/io/programs/impl/ClosePullRequestsProgram.kt
@@ -16,19 +16,19 @@ import pullrequestfactory.io.uis.ConsoleUI
 class ClosePullRequestsProgram(
         private val ui: UI,
         private val programArgs: ProgramArgs,
-        private val repoUrl: String) : Program {
+        private val repoUrl: String,
+        private val authToken: String) : Program {
 
     // TODO Add rate limit check
     override fun execute() {
         val candidate = programArgs.get_candidate()
-        val token = programArgs.get_github_basic_auth_token()
-
+        val token = programArgs.get_github_auth_token()
         val httpClient = KhttpClient(token)
         val rateLimitBefore = GithubAPIClient(httpClient).get_rate_limit()
         println("Rate rate limit before closing pull requests: $rateLimitBefore")
         println()
         val httpClientStats = KhttpClientStats(httpClient)
-        val branchesRepo = GithubHttpBranchesRepos(repoUrl, ui, httpClientStats)
+        val branchesRepo = GithubHttpBranchesRepos(repoUrl, ui, httpClientStats, authToken)
         val prRepo = GithubHttpPullRequestsRepo(repoUrl, token, ui, httpClientStats)
         val f = GithubPRFactory(
                 ConsoleUI(),