+++ src/main/kotlin/pullrequestfactory/io/repositories/GithubHttpBranchesRepo.kt
@@ -7,10 +7,11 @@ import pullrequestfactory.domain.branches.GithubBranchesRepo
 class GithubHttpBranchesRepo(
         private val repoUrl: String,
         private val response: Response,
-        private val httpClient: HttpClient) : GithubBranchesRepo {
+        private val httpClient: HttpClient,
+        private val authToken: String) : GithubBranchesRepo {
 
     override fun get_all_branches(): List<Branch> {
-        return GithubHttpReadRepo(httpClient).get_list(response, "$repoUrl/branches")
+        return GithubHttpReadRepo(httpClient).get_list(response, "$repoUrl/branches", authToken)
     }
 
 }