+++ src/main/kotlin/pullrequestfactory/io/repositories/GithubHttpPullRequestsRepo.kt
@@ -8,12 +8,12 @@ import pullrequestfactory.domain.uis.UI
 
 class GithubHttpPullRequestsRepo(
         private val repoPath: String,
-        private val basicAuthToken: String,
+        private val authToken: String,
         private val ui: UI,
         private val httpClient: HttpClient) : GithubPullRequestsRepo {
 
     override fun get_all_open_pull_requests(): List<GetPullRequest> {
-        return GithubHttpPullRequestsReadRepos(repoPath, ui, httpClient).get_all_open_pull_requests()
+        return GithubHttpPullRequestsReadRepos(repoPath, ui, httpClient, authToken).get_all_open_pull_requests()
     }
 
     override fun open_pull_request(pullRequest: PullRequest) {